requirements.txt
appdirs==1.4.3
asgiref==3.2.3
autopep8==1.5.2
beautiful==0.0.2
beautifulsoup4==4.9.0
bootstrap4==0.1.0
boto3==1.13.11
botocore==1.16.11
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
comtypes==1.1.7
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-appconf==1.0.4
django-bootstrap3==12.0.3
django-braces==1.14.0
django-heroku==0.3.1
django-imagekit==4.0.2
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
Faker==4.0.2
filelock==3.0.12
flake8==3.7.9
get==2019.4.13
gunicorn==20.0.4
idna==2.9
img2pdf==0.3.6
jmespath==0.10.0
lxml==4.5.0
mccabe==0.6.1
pilkit==2.0
Pillow==7.1.1
pipenv==2018.11.26
post==2019.4.13
psycopg2==2.8.5
public==2019.4.13
pycodestyle==2.5.0
pyflakes==2.1.1
pygame==1.9.6
PyMySQL==0.9.3
PyPDF2==1.26.0
pytesseract==0.3.4
python-dateutil==2.8.1
pyttsx3==2.87
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
s3transfer==0.3.3
selenium==3.141.0
six==1.14.0
soup==0.1.0
soupsieve==2.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.9
virtualenv==20.0.15
virtualenv-clone==0.5.4
whitenoise==5.0.1
